[NAV] Refine naming structure
A potentially-confusing aspect of navigation translation is that we have routeNames that are used under the hood for routing that are often the same english word as the visible label, but serve a different purpose. I landed on suffixing all visible labels with _LABEL, which will be used for translations, while routeNames will remain constant across languages.

NOTE: I refer to tab text as a 'label' and top title as 'title', but in certain cases a label is used as a title -- i.e. 'Applications' is technically a label when it appears on the bottom Tab Bar, but it's a title when you tap on the tab and it appears as the screen title up top. Usually, titles will be set dynamically based on the current Application/Gateway/Device the user is viewing, while labels are always consistent.

Also includes a naive hard-coded implementation of setting screen titles based on current Application/Gateway/Device.