[MQTT] Initial implmentation live data monitoring (#60)
Current for iOS only and is only on applications data view.

- Create bridge module to communicate with Native SDK
- Utils for converting base64 payloads to hex + permission check to ensure user has proper access key for mqtt connection. Both will be added to as we add more checks/features.
- Create ApplicationData screen with flat list including custom data list item component.

Downside of current implementation is one session at a time, meaning we can only listen to one application at a time. I don’t think that should be of any issue ATM since we our UI is not going to support that anyway.

We’ll need to follow up this PR to improve authorization and connection issues to the user.