[ProfileOverview] Add message on email change to confirm
I also decided to remove the logic that hid the tab bar on android when the keyboard is visible.
I did this originally because the main tab bar floats up and sits above the keyboard.
The solution ended up making it look worse causing a flicker and didn’t always work.