[FacebookSDK] Initial commit
Finally got an installation to show up for Android. I think the issue is that it doesn’t register sim installs and only registers a device the first time, not on every fresh install.
We should be able to fully test once we deploy the next version.

Production app id’s have already been added to our bitrise workflows.