[navigation] initial impl tab press reset if on active tab
If we are deep into a navigation stack and want to jump back to the home
screen, there is currently no way to do that except by pressing back
several times.

This change listens for presses on the currently active tab and will
reset the stack back to the home route.